Converting row counts in JSON to int
The previous push had the row counts stored as strings. If we want to do
anything automated with this in the future (?), it would be to our
benefit to store the row counts as integers as opposed to strings.

Updates the record structure and adds an `Atoi` for the row counts.

If the `Atoi` fails, we store -1.